Organize character feat collection classes and interfaces
