+++ src/module/actor/character/feats/index.ts
@@ -1,0 +1,3 @@
+export * from "./collection.ts";
+export * from "./group.ts";
+export * from "./types.ts";
