+++ src/module/item/feat/document.ts
@@ -1,6 +1,6 @@
 import type { ActorPF2e } from "@actor";
 import { ClassDCData } from "@actor/character/data.ts";
-import type { FeatGroup } from "@actor/character/feats.ts";
+import type { FeatGroup } from "@actor/character/feats/index.ts";
 import type { SenseData } from "@actor/creature/index.ts";
 import { ItemPF2e, type HeritagePF2e } from "@item";
 import { getActionCostRollOptions, normalizeActionChangeData, processSanctification } from "@item/ability/helpers.ts";
